third_party/WebKit/LayoutTests/platform/win-xp/fast/dom/Range/getBoundingClientRect-caret-expected.txt

hello
Test getBoundingClientRect of caret for collapsed range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.left is 22
PASS rect.top is 8
PASS rect.width is 0
PASS rect.height is 19
PASS successfullyParsed is true

TEST COMPLETE
